Remove feature that extracts Makefile.PL & Build.PL
Because this is unsafe (we are running untrusted code). Note that MetaCPAN also
does not extract dist metadata from Makefile.PL and Build.PL for the same
reason.

Also, it has been pointed out to me that the current way is not proper
("idiomatic"). The more proper way is to let perl run Makefile.PL and Build.PL
to generate Makefile/Build as well as MYMETA.{json,yml}. Then we read MYMETA.
Perhaps this feature will come back as an optional plugin later, should there be
any demand. BTW, Dist::Requires does this kind of functionality.

Also remove the feature that detects *.meta files, because these files are not
present in a mini CPAN mirror.

Upgrade schema to v3 to add has_* columns.